+++ packages/auth/src/api/account_management/email_and_password.ts
@@ -25,6 +25,7 @@ import {
 } from '../index';
 import { IdTokenResponse } from '../../model/id_token';
 import { MfaEnrollment } from './mfa';
+import { SignUpRequest, SignUpResponse } from '../authentication/sign_up';
 
 export interface ResetPasswordRequest {
   oobCode: string;
@@ -67,6 +68,20 @@ export async function updateEmailPassword(
     UpdateEmailPasswordRequest,
     UpdateEmailPasswordResponse
   >(auth, HttpMethod.POST, Endpoint.SET_ACCOUNT_INFO, request);
+}
+
+// Used for linking an email/password account to an existing idToken. Uses the same request/response
+// format as updateEmailPassword.
+export async function linkEmailPassword(
+  auth: Auth,
+  request: SignUpRequest
+): Promise<SignUpResponse> {
+  return _performApiRequest<SignUpRequest, SignUpResponse>(
+    auth,
+    HttpMethod.POST,
+    Endpoint.SIGN_UP,
+    request
+  );
 }
 
 export interface ApplyActionCodeRequest {