+++ packages/app-compat/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @firebase/app-compat
 
+## 0.2.23
+
+### Patch Changes
+
+- Updated dependencies [[`5c7fa8491`](https://github.com/firebase/firebase-js-sdk/commit/5c7fa84912ac8a9652b82ebf88eb483dd44977a8)]:
+  - @firebase/app@0.9.23
+
 ## 0.2.22
 
 ### Patch Changes